Provide a simple example of object reflection in Java
In order to use reflection in Java, you must include the necessary
classes from the java.lang.reflect package since that contains all the
necessary classes for implementing reflection. The one exception to this
is when retrieving class metadata – for which you would use the
java.lang.Class class, which is not part of the java.lang.reflect
package.

Let’s look at an example of Java reflection to help clarify exactly what
it does and when it can be useful.